python/data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "mysql-connector>=2.2.9",
]
